Refactor: 친구 목록 저장 시 권한 설정
카카오톡 친구 목록을 저장할 때 권한을 BETA로 설정하도록 수정
