+++ src/main/java/supernova/whokie/friend/infrastructure/repository/FriendRepository.java
@@ -1,4 +1,4 @@
-package supernova.whokie.friend.repository;
+package supernova.whokie.friend.infrastructure.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
