+++ src/main/java/supernova/whokie/friend/service/FriendService.java
@@ -5,9 +5,11 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import supernova.whokie.friend.Friend;
 import supernova.whokie.friend.controller.dto.FriendRequest;
-import supernova.whokie.friend.repository.FriendRepository;
-import supernova.whokie.friend.service.dto.KakaoDto;
+import supernova.whokie.friend.infrastructure.apiCaller.FriendKakaoApiCaller;
+import supernova.whokie.friend.infrastructure.repository.FriendRepository;
+import supernova.whokie.friend.infrastructure.apiCaller.dto.KakaoDto;
 import supernova.whokie.global.exception.EntityNotFoundException;
+import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
 import supernova.whokie.user.repository.UserRepository;
 
@@ -51,6 +53,7 @@ public class FriendService {
                         .name(profile.profileNickname())
                         .kakaoCode(profile.uuid())
                         .imageUrl(profile.profileThumbnailImage())
+                        .role(Role.BETA)
                         .build())
                 .toList();
     }
@@ -81,7 +84,10 @@ public class FriendService {
 
         return friendUsers.stream()
                 .filter(friend -> !existingFriendIds.contains(friend.getId()))
-                .map(friend -> Friend.builder().hostUser(host).friendUser(friend).build())
+                .map(friend -> Friend.builder()
+                        .hostUser(host)
+                        .friendUser(friend)
+                        .build())
                 .toList();
     }
 }
